requirements.txt
# For Knowledge Base stuff
openai-whisper
langchain
chromadb
InstructorEmbedding
sentence-transformers

# GPT
faiss-cpu
huggingface_hub
transformers
protobuf
auto-gptq


# Utility library
python-dotenv
pathlib
urllib3
accelerate
bitsandbytes
click